accumlate into test cases

+++ code/library/spline/src/sparse_utilities.cpp
@@ -2,7 +2,7 @@
 
 namespace reprojection::spline {
 
-Eigen::SparseMatrix<double> DiagonalSparseMatrix(MatrixXd const& block, int const stride, size_t const count) {
+Eigen::SparseMatrix<double> DiagonalSparseMatrix(MatrixXd const& block, size_t const stride, size_t const count) {
     if (not(block.rows() == block.cols())) {
         throw std::runtime_error("Only accepts square blocks");
     }
@@ -11,7 +11,7 @@ Eigen::SparseMatrix<double> DiagonalSparseMatrix(MatrixXd const& block, int cons
     triplets.reserve(block.size() * count);
 
     for (size_t n{0}; n < count; n++) {
-        int const start_index{static_cast<int>(n * stride)};
+        size_t const start_index{n * stride};
 
         for (Eigen::Index row{0}; row < block.rows(); row++) {
             for (Eigen::Index col{0}; col < block.cols(); col++) {
