+++ code/library/spline/src/sparse_utilities.hpp
@@ -7,6 +7,6 @@
 namespace reprojection::spline {
 
 // See section "Filling a sparse matrix" - https://libeigen.gitlab.io/eigen/docs-nightly/group__TutorialSparse.html
-Eigen::SparseMatrix<double> DiagonalSparseMatrix(MatrixXd const& block, int const stride, size_t const count);
+Eigen::SparseMatrix<double> DiagonalSparseMatrix(MatrixXd const& block, size_t const stride, size_t const count);
 
 }  // namespace reprojection::spline
